chore: rename existing 'NetworkMonitorAgent' to 'NodeStressTester'
